Cart Manager + Sell Button
Items, when clicked on, now get added to a list, and clicking them again removes them from the list. Added a "Clear Items" button, which quickly removes every item from the cart. Also added a "Sell" button, which removes the selected items from the scene. Plan to add a restart feature when all items are sold.