Customer, Font, New Camera Movement
The customer will leave the shop after a sale and come back a few seconds later. Implemented new camera movement to jump to the store front when selling items and back to Harlow when the customer comes back.
Updated scoring with an Unsold Items list; each item requested that is not sold deducts points from the player.